+++ CHANGES.md
@@ -7,6 +7,8 @@
 -   Enhancement [FRD]: Add all FRD models, including API pages, serializers and filters
 -   Enhancement [GAR]: Improve GAR-import to include all data, instead of just the main information of GAR
 -   Enhancement [FRD]: Improve FRD-import to include all data, instead of just the main dossier.
+-   Bugfix [GAR]: Correct the urls for the API of the GAR
+-   Bugfix [GMN]: Correct Move MonitoringTube Request
 
 
 ## 1.68 (2026-04-10)